LibWebView+WebContent: Move compositor client code into LibWebView
The CompositorConnection and the CompositorHost implementation driving
it lived in Services/WebContent, which only WebContent can link. The
WebWorker process is about to need the same plumbing so OffscreenCanvas
contexts created in workers can stream canvas commands to the
Compositor process.

Move CompositorConnection into LibWebView unchanged, and pull the
reusable part of WebContentCompositorHost into a new
WebView::CompositorHostBase whose subclasses supply the connection.
WebContent keeps a thin subclass that routes context-destruction
notifications back to its client connection. LibWebView now links
LibMedia because the moved connection encodes Media::VideoFrame for
the video frame update messages. No behavior change.
