+++ Libraries/LibWebView/CMakeLists.txt
@@ -10,6 +10,8 @@ set(SOURCES
     CanonicalTraversable.cpp
     BrowserProcess.cpp
     CompositorClient.cpp
+    CompositorConnection.cpp
+    CompositorHostBase.cpp
     ConsoleOutput.cpp
     CookieJar.cpp
     DictionaryLookup.cpp
@@ -102,7 +104,7 @@ compile_ipc(${LADYBIRD_SOURCE_DIR}/Services/WebContent/WebUIClient.ipc ${CMAKE_B
 compile_ipc(${LADYBIRD_SOURCE_DIR}/Services/WebContent/WebUIServer.ipc ${CMAKE_BINARY_DIR}/Services/WebContent/WebUIServerEndpoint.h)
 
 ladybird_lib(LibWebView webview EXPLICIT_SYMBOL_EXPORT)
-target_link_libraries(LibWebView PRIVATE LibCore LibCrypto LibDatabase LibDevTools LibFileSystem LibGfx LibHTTP LibImageDecoderClient LibIPC LibRequests LibJS LibWeb LibUnicode LibURL LibSync LibSyntax LibTextCodec LibThreading)
+target_link_libraries(LibWebView PRIVATE LibCore LibCrypto LibDatabase LibDevTools LibFileSystem LibGfx LibHTTP LibImageDecoderClient LibIPC LibRequests LibJS LibMedia LibWeb LibUnicode LibURL LibSync LibSyntax LibTextCodec LibThreading)
 
 # Third-party
 if (HAS_FONTCONFIG)
