+++ Services/WebContent/ConnectionFromClient.h
@@ -55,7 +55,7 @@ public:
 
     PageHost& page_host() { return *m_page_host; }
     PageHost const& page_host() const { return *m_page_host; }
-    CompositorConnection* compositor_process_connection() const;
+    WebView::CompositorConnection* compositor_process_connection() const;
     void did_destroy_compositor_context(Web::Compositor::CompositorContextId);
 
     Function<void(IPC::TransportHandle const&)> on_request_server_connection;
@@ -234,7 +234,7 @@ private:
 
     virtual void exit_fullscreen(u64 page_id) override;
 
-    RefPtr<CompositorConnection> m_compositor_connection;
+    RefPtr<WebView::CompositorConnection> m_compositor_connection;
     NonnullOwnPtr<PageHost> m_page_host;
 
     HashMap<int, Web::FileRequest> m_requested_files {};
