+++ Services/WebContent/Forward.h
@@ -8,7 +8,6 @@
 
 namespace WebContent {
 
-class CompositorConnection;
 class ConnectionFromClient;
 class ConsoleGlobalEnvironmentExtensions;
 class DevToolsConsoleClient;
